requirements.txt
 
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
google-cloud-bigquery>=3.4.0
google-auth>=2.0.0
pyarrow>=10.0.0
wordcloud>=1.9.0
scikit-learn>=1.3.0
networkx>=3.1